+++ views/partials/header.handlebars
@@ -10,8 +10,6 @@
         </div>
         <div class="navbar-collapse collapse navbar-responsive-collapse">
             <ul class="nav navbar-nav">
-                <li class=""><a href="/#">Link</a>
-                </li>
                 <li class="" id="userinfo" user="{{currentuser.name}}" uid="{{currentuser._id}}"></li>
             </ul>
             <ul class="nav navbar-nav navbar-right">
